refactor: move namespace local\renderables to output
