+++ classes/output/course_event_slot_table.php
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace block_coursefeedback\local\renderables;
+namespace block_coursefeedback\output;
 
 use block_coursefeedback\local\course_feedback_data;
 use block_coursefeedback\local\persistent\eventtype;
