+++ classes/output/survey_execution_period.php
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace block_coursefeedback\local\renderables;
+namespace block_coursefeedback\output;
 
 use block_coursefeedback\local\persistent\survey_execution;
 use core\output\inplace_editable;
